feat(fetch): pure-ESM web testing
- Add `web-test-runner` and associated dependencies
- Add invocation point for pure ESM tests
