+++ package.json
@@ -108,8 +108,8 @@
   "type": "module",
   "types": "index.d.ts",
   "scripts": {
-    "test": "npm run test:eslint && npm run test:mocha && npm run test:karma && npm run test:dtslint && npm run test:exports",
-    "test:all": "npm run test:node && npm run test:deno && npm run test:workers",
+    "test": "npm run test:eslint && npm run test:mocha && HEADLESS=true npm run test:browser && npm run test:dtslint && npm run test:exports",
+    "test:all": "npm run test:node && npm run test:deno && npm run test:workers && HEADLESS=true npm run test:browser:modern",
     "test:node": "npm run test",
     "test:eslint": "node bin/ssl_hotfix.js eslint lib/**/*.js",
     "test:dtslint": "dtslint --localTs node_modules/typescript/lib",
@@ -117,8 +117,10 @@
     "test:deno": "deno test test/fetch/deno/fetch.deno.spec.js",
     "test:exports": "node bin/ssl_hotfix.js mocha test/module/test.js --timeout 30000 --exit",
     "test:workers": "echo \"Workers test not implemented yet.\"",
-    "test:karma": "node bin/ssl_hotfix.js cross-env LISTEN_ADDR=:: karma start karma.conf.cjs --single-run",
+    "test:karma": "npm run test:browser",
     "test:karma:firefox": "node bin/ssl_hotfix.js cross-env LISTEN_ADDR=:: Browsers=Firefox karma start karma.conf.cjs --single-run",
+    "test:browser": "node bin/ssl_hotfix.js cross-env LISTEN_ADDR=:: karma start karma.conf.cjs --single-run",
+    "test:browser:modern": "node bin/ssl_hotfix.js wtr test/fetch/generic/harness.html --playwright --browsers chromium firefox webkit",
     "test:karma:server": "node bin/ssl_hotfix.js cross-env karma start karma.conf.cjs",
     "test:build:version": "node ./bin/check-build-version.js",
     "start": "node ./sandbox/server.js",
@@ -165,6 +167,7 @@
     "@babel/preset-env": "^7.23.9",
     "@commitlint/cli": "^17.8.1",
     "@commitlint/config-conventional": "^17.8.1",
+    "@esm-bundle/chai": "^4.3.4-fix.0",
     "@release-it/conventional-changelog": "^5.1.1",
     "@rollup/plugin-alias": "^4.0.2",
     "@rollup/plugin-babel": "^5.3.1",
@@ -172,6 +175,8 @@
     "@rollup/plugin-json": "^4.1.0",
     "@rollup/plugin-multi-entry": "^4.1.0",
     "@rollup/plugin-node-resolve": "^9.0.0",
+    "@web/test-runner": "^0.15.0",
+    "@web/test-runner-playwright": "^0.9.0",
     "abortcontroller-polyfill": "^1.7.5",
     "auto-changelog": "^2.4.0",
     "body-parser": "^1.20.2",
