+++ web-test-runner.config.mjs
@@ -1,0 +1,5 @@
+export default {
+  concurrency: 10,
+  coverage: true,
+  nodeResolve: true,
+};
